Add files via upload
This is a sample scene of the new textures added to the colosseum. Also been working on a path that leads to a teleporter which will lead to a Teleporting island. But need to work on the physics of it. Also added a little terrain to the actual colosseum battleground and imported the benches to the stands and going to add doors with animations in the next few days.